Using commands/queries for UniqueAccount

+++ WebAPI/Controllers/UniqueAccountController.cs
@@ -1,5 +1,8 @@
 using Domain.Aggregate;
+using Domain.Command;
+using Domain.DTO;
 using Domain.Service;
+using Infra.DataFromMongo.Entity;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -19,51 +22,75 @@ public class UniqueAccountController : ControllerBase
         _uniqueAccountService = uniqueAccountService;
     }
 
-    // GET: api/uniqueaccounts
+    // GET: uniqueaccounts
     [HttpGet]
     public async Task<IActionResult> GetAllUniqueAccounts()
     {
         var uniqueAccounts = await _uniqueAccountService.GetAllUniqueAccountsAsync();
-        return Ok(uniqueAccounts);
+        if (uniqueAccounts == null)
+        {
+            return NotFound();
+        }
+        return Ok(UniqueAccountResultDTO.GenerateListFromAggregates(uniqueAccounts));
     }
 
-    // GET: api/uniqueaccounts/{id}
+    // GET: uniqueaccounts/{id}
     [HttpGet("{id}")]
-    public async Task<IActionResult> GetUniqueAccountById(string id)
+    public async Task<IActionResult> GetUniqueAccountById([FromRoute] string id)
     {
+        if (string.IsNullOrWhiteSpace(id)) return BadRequest();
         var uniqueAccount = await _uniqueAccountService.GetUniqueAccountByIdAsync(id);
         if (uniqueAccount == null)
         {
             return NotFound();
         }
-        return Ok(uniqueAccount);
+        return Ok(new UniqueAccountResultDTO(uniqueAccount));
     }
 
-    // POST: api/uniqueaccounts
+    // POST: uniqueaccounts
     [HttpPost]
-    public async Task<IActionResult> CreateUniqueAccount([FromBody] UniqueAccountAggregate uniqueAccount)
+    public async Task<IActionResult> CreateUniqueAccount([FromBody] AddUniqueAccountCommand uniqueAccount)
     {
-        var createdUniqueAccount = await _uniqueAccountService.CreateUniqueAccountAsync(uniqueAccount);
+        if (uniqueAccount == null) return BadRequest();
+        var createdUniqueAccount = await _uniqueAccountService.CreateUniqueAccountAsync(new UniqueAccountAggregate
+        {
+            DisplayName = uniqueAccount.DisplayName,
+            Email = uniqueAccount.Email,
+            IsAdmin = uniqueAccount.IsAdmin,
+            Name = uniqueAccount.Name,
+            Password = uniqueAccount.Password,
+        });
         //return CreatedAtAction(nameof(GetUniqueAccountById), new { success = createdUniqueAccount }, createdUniqueAccount);
         return Ok(new { success = createdUniqueAccount });
     }
 
-    // PUT: api/uniqueaccounts/{id}
+    // PUT: uniqueaccounts/{id}
     [HttpPut("{id}")]
-    public async Task<IActionResult> UpdateUniqueAccount(string id, [FromBody] UniqueAccountAggregate uniqueAccount)
+    public async Task<IActionResult> UpdateUniqueAccount([FromRoute] string id, [FromBody] UpdUniqueAccountCommand uniqueAccount)
     {
-        var updatedUniqueAccount = await _uniqueAccountService.UpdateUniqueAccountAsync(uniqueAccount);
+        if (uniqueAccount == null) return BadRequest();
+        if (string.IsNullOrWhiteSpace(id)) return BadRequest();
+        var updatedUniqueAccount = await _uniqueAccountService.UpdateUniqueAccountAsync(new UniqueAccountAggregate
+        {
+            Id = id,
+            DisplayName = uniqueAccount.DisplayName,
+            Email = uniqueAccount.Email,
+            IsAdmin = uniqueAccount.IsAdmin,
+            Name = uniqueAccount.Name,
+            Password = uniqueAccount.Password
+        });
         if (!updatedUniqueAccount)
         {
             return NotFound();
         }
-        return Ok(updatedUniqueAccount);
+        return Ok(new { success = updatedUniqueAccount });
     }
 
-    // DELETE: api/uniqueaccounts/{id}
+    // DELETE: uniqueaccounts/{id}
     [HttpDelete("{id}")]
-    public async Task<IActionResult> DeleteUniqueAccount(string id)
+    public async Task<IActionResult> DeleteUniqueAccount([FromRoute] string id)
     {
+        if (string.IsNullOrWhiteSpace(id)) return BadRequest();
         var isDeleted = await _uniqueAccountService.DeleteUniqueAccountAsync(id);
         if (!isDeleted)
         {
